rename __session_wrapper to wrapper for PEP8 module naming conventions
